+++ backend/Api_Backend/src/main/java/com/example/restservice/EmergencyRecord.java
@@ -1,3 +1,3 @@
 package com.example.restservice;
 
-public record EmergencyRecord(String emergency, String injury, int time, float latitude, float longitude) {}
+public record EmergencyRecord(String emergency, String injury, Long time, float latitude, float longitude) {}
